+++ AegisTracker/View/ExpenseDateView.swift
@@ -90,7 +90,7 @@ struct ExpenseDateView: View {
                             .fontWeight(.bold)
                             .fontDesign(.rounded)
                     }
-                    FinanceMonthChart(data: monthExpenses.map(Expense.toFinanceData), year: year, month: month)
+                    FinanceMonthLineChart(data: monthExpenses.map(Expense.toFinanceData), year: year, month: month)
                         .frame(height: 140)
                 }
             } header: {