+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Expense tags: expenses can now have multiple 'tags'
 - Method of quickly adding tags across multiple expenses
 - Method of creating a new expense/income from existing instance (template)
+- Updated month chart to be line chart with old expenses
+
+### Removed
+
+- Animation when switching date periods (buggy, need to rework)
 
 ## [0.1.0] - 2025-01-01